api/requirements.txt
autopep8==2.0.2
bcrypt==4.0.1
blinker==1.6.2
click==8.1.3
distlib==0.3.6
dnspython==2.3.0
filelock==3.12.0
Flask==2.2.5
Flask-Cors==3.0.10
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
platformdirs==3.5.1
pycodestyle==2.10.0
pymongo==4.3.3
python-dotenv==1.0.0
six==1.16.0
SQLAlchemy==2.0.15
typing_extensions==4.5.0
virtualenv==20.23.0
Werkzeug==2.3.4